refactor: move scene editor converters to dedicated resource dictionary
