+++ WolvenKit/Views/Documents/CombinedSceneView.xaml.cs
@@ -21,26 +21,6 @@ using Splat;
 namespace WolvenKit.Views.Documents
 {
     /// <summary>
-    /// Converter to map a path type to connection thickness.
-    /// </summary>
-    public sealed class PathTypeToThicknessConverter : IValueConverter
-    {
-        public double ThinThickness { get; set; } = 1.5;
-        public double ThickThickness { get; set; } = 3.0;
-
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            // PathType 1 is Dead End, 2 is Live Path
-            return (value is int pathType && pathType == 1) ? ThinThickness : ThickThickness;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    /// <summary>
     /// Interaktionslogik für CombinedSceneView.xaml
     /// </summary>
     public partial class CombinedSceneView : UserControl
@@ -429,8 +409,6 @@ namespace WolvenKit.Views.Documents
             
             timer.Start();
         }
-
-
 
         /// <summary>
         /// Handles keyboard shortcuts for node deletion and navigation
